user_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib
python-jose
pytest
datetime